backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
numpy==1.26.4
ccxt==4.2.15
python-dotenv==1.0.1
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==6.0.1